수정 초기버전 검사용
