+++ src/components/map/KakaoMapContainer.module.scss
@@ -1,0 +1,11 @@
+.map-wrap {
+  width: 100vw;
+  height: 100vh;
+}
+
+
+// .map-side{
+//   z-index: 1;
+//   display: flex;
+//   float: left;
+// }
